third_party/WebKit/LayoutTests/fast/js/Promise-catch-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInInit === firstPromise is false
PASS thisInInit === window is true
PASS firstPromise instanceof Promise is true
PASS secondPromise instanceof Promise is true
PASS firstPromise.catch(null) threw exception TypeError: onRejected must be a function or undefined.
PASS firstPromise.catch(37) threw exception TypeError: onRejected must be a function or undefined.
PASS thisInFulfillCallback === firstPromise is false
PASS thisInFulfillCallback === secondPromise is false
PASS thisInFulfillCallback === window is true
PASS result is "hello"
PASS result is "bye"
PASS fulfilled
PASS successfullyParsed is true

TEST COMPLETE
